Implementes the option of implementing inline functions as C macros including one-line macros

+++ ports/Microchip/PIC16F-generic/api/pin/setup.c.v
@@ -10,13 +10,12 @@ module pin
 @[inline]
 pub fn setup(name u8, mode u8) {
 	unsafe { 
-		// addr = &C.TRISA + (name >> 3)  // &C.TRISA + (name / 8)		// base address plus shift
 		if mode == input { // as input (1)
-			*(&C.TRISA + (name >> 3)) |= (u8(0x01) << (name - ((name >> 3) << 3)))
-			// |=  (u8(0x01) << (name % 8))
+			*(&C.TRISA + (name >> 3)) |= (0x01 << (name - ((name >> 3) << 3)))
+		//  *(&C.TRISA + (name / 8))  |= (0x01 << (name % 8))
 		} else { // as output (0)
-			*(&C.TRISA + (name >> 3)) &= ~(u8(0x01) << (name - ((name >> 3) << 3)))
-			// &= ~(u8(0x01) << (name % 8))
+			*(&C.TRISA + (name >> 3)) &= ~(0x01 << (name - ((name >> 3) << 3)))	
+		//  *(&C.TRISA + (name / 8))  &= ~(0x01 << (name % 8))
 		}
 	}
 }
