+++ ports/Microchip/PIC16F-generic/api/pin/toggle.c.v
@@ -10,6 +10,6 @@ module pin
 @[inline]
 pub fn toggle(name u8) {
 	unsafe {
-		*(&C.PORTA + u8(name >> 3))  ^=  u8(0x01) << (name - ((name >> 3) << 3))
-	}// *(&C.PORTA + u8(name / 8))  ^=  0x01 << (name % 8)
+		*(&C.PORTA + (name >> 3))  ^=  0x01 << (name - ((name >> 3) << 3))
+	}// *(&C.PORTA + (name / 8))   ^=  0x01 << (name % 8)
 }
